requirements.txt
contourpy==1.3.3
customtkinter==5.2.2
cycler==0.12.1
darkdetect==0.8.0
dotenv==0.9.9
ecdsa==0.19.1
et_xmlfile==2.0.0
fonttools==4.59.2
kiwisolver==1.4.9
MarkupSafe==3.0.2
matplotlib==3.10.6
numpy==2.3.3
openpyxl==3.1.5
packaging==25.0
pandas==2.3.2
pillow==11.3.0
pyglet==2.1.8
pyparsing==3.2.4
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-http-client==3.3.7
pytz==2025.2
sendgrid==6.12.4
six==1.17.0
tzdata==2025.2
Werkzeug==3.1.3
